w_min :int --> float

+++ Bacteria.h
@@ -34,15 +34,15 @@ class Bacteria {
         // ===================================================================
         //                           Attributes
         // ===================================================================
-        static int w_min; // Minimum allowed fitness value
+        static float w_min; // Minimum allowed fitness value
         static float p_mut;
         static float p_death;
 
         float w; // Bacteria's fitness
         int genotype; // Bacteria's genotype
-        float A;
-        float B;
-        float C;
+        float A; // Internal concentration of Glucose
+        float B; // Internal concentration of Acetate
+        float C; // Internal concentration of Ethanol
 
 };
